feat(item): get-all-item store & handler
- update item float32 --> float64
- item-store - select all from db
- item-handler - return all items

+++ divine-server/types/item/type.go
@@ -3,8 +3,8 @@ package types
 type Item struct {
 	ID            int
 	Name          string
-	CurrentPrice  float32
-	PreviousPrice float32
+	CurrentPrice  float64
+	PreviousPrice float64
 }
 
 type ItemRequest struct {
